1. Remove mixin class by consolidating metadata inside the AbstractPotential class
